requirements.txt
absl-py==0.13.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
autoattack @ git+https://github.com/fra31/auto-attack.git@9b264b52bb65c373373727f532865a5551ab9c02
cachetools==4.2.2
certifi==2021.5.30
chardet==3.0.4
charset-normalizer==2.0.3
click==8.0.1
configparser==5.0.2
docker-pycreds==0.4.0
fsspec==2021.7.0
future==0.18.2
gitdb==4.0.7
GitPython==3.1.18
google-auth==1.34.0
google-auth-oauthlib==0.4.4
grpcio==1.39.0
idna==2.10
importlib-metadata==4.6.1
Jinja2==2.11.3
lightning-bolts==0.3.4
Markdown==3.3.4
MarkupSafe==2.0.1
multidict==5.1.0
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.3.56
packaging==21.0
pandas==1.1.5
pathtools==0.1.2
Pillow==8.3.1
promise==2.3
protobuf==3.17.3
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.1
pyparsing==2.4.7
python-dateutil==2.8.2
pytorch-lightning==1.4.0
pytz==2021.1
PyYAML==5.4.1
requests==2.23.0
requests-oauthlib==1.3.0
robustbench @ git+https://github.com/RobustBench/robustbench.git@4af5687974145a6e1fd27ab4a4183ae7365f267e
rsa==4.7.2
sentry-sdk==1.3.1
shortuuid==1.0.1
six==1.16.0
smmap==4.0.0
subprocess32==3.5.4
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
torch==1.9.0+cu111
torchaudio==0.9.0
torchmetrics==0.4.1
torchvision==0.10.0+cu111
tqdm==4.56.2
typing-extensions==3.10.0.0
urllib3==1.25.11
wandb==0.11.0
Werkzeug==2.0.1
yarl==1.6.3
zipp==3.5.0

# package location
--find-links https://download.pytorch.org/whl/torch_stable.html